更新网络结构

+++ src/network/sampling.py
@@ -35,10 +35,10 @@ class Sampler(nn.Module):
             #nn.ReLU(),
         )
 
-        self.mask = torch.zeros(self.max_len, dtype=torch.float32, device=device)
+        self.mask = torch.zeros(self.max_len, device=device)
         self.mask[::self.attributes_num] = 1
         self.mask = self.mask.unsqueeze(0).expand(self.max_len, -1)
-        self.inv_mask = torch.ones(self.max_len, dtype=torch.float32, device=device)
+        self.inv_mask = torch.ones(self.max_len, device=device)
         self.inv_mask[::self.attributes_num] = 0
         self.inv_mask = self.inv_mask.unsqueeze(0).expand(self.max_len, -1)
